+++ packages/contracts-bridge/contracts/test/TokenRegistry.t.sol
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: MIT OR Apache-2.0
+pragma solidity 0.7.6;
+
+import {BridgeTest} from "./utils/BridgeTest.sol";
+
+contract TokenRegistryTest is BridgeTest {
+    function setUp() public override {
+        super.setUp();
+    }
+}