+++ packages/indexer/src/api/index.ts
@@ -7,36 +7,49 @@ import {
   ApolloServerPluginLandingPageGraphQLPlayground,
 } from 'apollo-server-core';
 import http from 'http';
-import {
-  getDB,
-  getGraphqlSchema,
-  getMetricsMiddleware,
-  initSentry,
-} from './utils';
+import promBundle from 'express-prom-bundle';
+import { register } from 'prom-client';
+import { getGraphqlSchema, initSentry } from './utils';
 import * as Sentry from '@sentry/node';
 import { isProduction, port } from '../config';
-import router from './routes';
+import { getRouter } from './routes';
+import { DB } from '../core/db';
+import Logger from 'bunyan';
+import { prefix } from '../core/metrics';
 
-const app = express();
+export async function run(db: DB, logger: Logger) {
+  const app = express();
 
-if (isProduction) {
-  initSentry(app);
-  // RequestHandler creates a separate execution context using domains, so that every
-  // transaction/span/breadcrumb is attached to its own Hub instance
-  app.use(Sentry.Handlers.requestHandler());
-  // TracingHandler creates a trace for every incoming request
-  app.use(Sentry.Handlers.tracingHandler());
-}
+  if (isProduction) {
+    initSentry(app);
+    // RequestHandler creates a separate execution context using domains, so that every
+    // transaction/span/breadcrumb is attached to its own Hub instance
+    app.use(Sentry.Handlers.requestHandler());
+    // TracingHandler creates a trace for every incoming request
+    app.use(Sentry.Handlers.tracingHandler());
+  }
 
-app.use(cors());
-app.disable('x-powered-by');
+  app.use(cors());
+  app.disable('x-powered-by');
 
-// add all the routes
-app.use('/', router);
+  app.use(
+    promBundle({
+      httpDurationMetricName: prefix + '_api',
+      buckets: [0.1, 0.3, 0.6, 1, 1.5, 2.5, 5],
+      includeMethod: true,
+      includePath: true,
+      promRegistry: register,
+    }),
+  );
 
-// TODO: consider supporting top level await in monorepo
-// (eg: module: es2022, target: >es2017 in the tsconfig)
-(async () => {
+  const router = await getRouter(db, logger);
+
+  // add all the routes
+  app.use('/', router);
+
+  // TODO: consider supporting top level await in monorepo
+  // (eg: module: es2022, target: >es2017 in the tsconfig)
+
   const httpServer = http.createServer(app);
 
   const graphqlServer = new ApolloServer({
@@ -48,8 +61,9 @@ app.use('/', router);
     ],
     introspection: true,
     context: {
-      prisma: (await getDB()).client,
+      prisma: db.client,
     },
+    cache: 'bounded',
   });
 
   if (isProduction) {
@@ -57,10 +71,8 @@ app.use('/', router);
     app.use(Sentry.Handlers.errorHandler());
   }
 
-  app.use(getMetricsMiddleware());
-
   await graphqlServer.start();
   graphqlServer.applyMiddleware({ app });
 
   httpServer.listen({ port });
-})();
+}